dbi and dbd: expose DATETIME fields as a struct tm instead
Dates before the epoch are not (always) representable in time_t. In
particular, the Microsoft C runtime's mktime function rejects any
dates before 1970, which is different from glibc which tries to
produce negative time_t values.

The sensible thing to do is to expose the broken-down date, for
example in a struct tm.